+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 FEATURES:
 * mdb: support Redis 6.0 in `yandex_mdb_redis_cluster`
 
+FEATURES:
+* **New Data Source:** `yandex_client_config`
+
 ENHANCEMENTS:
 * mdb: add `role` attribute to `host` entity in `yandex_mdb_postgresql_cluster` resource and data source